feat: UserSchema

+++ src/services/category.services.ts
@@ -1,12 +1,20 @@
 import { prisma } from "../database/prisma";
+import { AppError } from "../errors/appError";
 import { TCategory, TCategoryCreate } from "../schemas/categories.schema";
 
 export class CategoryServices {
-    public async create(body: TCategoryCreate): Promise<TCategory> {
-        return await prisma.category.create({ data: body });
+    public async create(body: TCategoryCreate, userId?: number): Promise<TCategory> {
+        if(!userId){
+            throw new AppError(403, "This user is not the category owner")
+        }
+        return await prisma.category.create({ data: {...body, userId: userId} });
     }
 
-    public async delete(categoryId: number): Promise<void> {
-        await prisma.category.delete({ where: { id: categoryId } });
+    public async delete(categoryId: number, userId?: number): Promise<void> {        
+        if(!userId){
+            throw new AppError(403, "This user is not the category owner")
+        }     
+
+        await prisma.category.delete({ where: { id: categoryId , userId:userId}});
     }
 }
